@@ -1,42 +1,33 @@
 package de.sremer.crawlicious.model;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.Date;
 
 @Entity
+@Getter
+@Setter
+@NoArgsConstructor
 public class PasswordResetToken {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Getter
-    @Setter
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Getter
-    @Setter
     private String token;
 
     @OneToOne(targetEntity = User.class, fetch = FetchType.EAGER)
     @JoinColumn(nullable = false, name = "user_id")
-    @Getter
-    @Setter
     private User user;
 
-    @Getter
-    @Setter
     private boolean valid = true;
 
-    @Getter
-    @Setter
     private Date expiryDate;
-
-    public PasswordResetToken() {
-    }
 
     public PasswordResetToken(String token, User user) {
         this.token = token;
