@@ -1,83 +1,82 @@
 package de.sremer.crawlicious.configuration;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.switchuser.SwitchUserFilter;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
+@RequiredArgsConstructor
 @EnableWebSecurity
-public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+public class SecurityConfiguration {
 
-    @Autowired
-    private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    @Autowired
     @Qualifier("userService")
-    private UserDetailsService userDetailsService;
-
-    @Value("${spring.queries.users-query}")
-    private String usersQuery;
-
-    @Value("${spring.queries.roles-query}")
-    private String rolesQuery;
+    private final UserDetailsService userDetailsService;
 
     @Value("${rememberme-key}")
     private String rememberMeKey;
 
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth)
-            throws Exception {
-        auth
-                .userDetailsService(userDetailsService)
-                .passwordEncoder(bCryptPasswordEncoder);
-    }
+    @Bean
+    public SecurityFilterChain configureSecurity(HttpSecurity http) throws Exception {
+        return http
+                .rememberMe(rememberMeConfigurer -> rememberMeConfigurer
+                        .rememberMeCookieName("woofles-remember-me")
+                        .key(rememberMeKey)
+                        .tokenValiditySeconds(60 * 60 * 24 * 365))
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-                .rememberMe().rememberMeCookieName("woofles-remember-me").tokenValiditySeconds(60 * 60 * 24 * 365).key(rememberMeKey)
-
-                .and()
                 .addFilter(switchUserFilter())
+                .csrf(AbstractHttpConfigurer::disable)
+                .formLogin(loginCustomizer -> loginCustomizer
+                        .loginPage("/login").failureUrl("/?login=error")
+                        .defaultSuccessUrl("/profile")
+                        .usernameParameter("email")
+                        .passwordParameter("password"))
 
-                .authorizeRequests()
-                .antMatchers("/", "/login", "/profile/**", "/test").permitAll()
-                .antMatchers("/imprint").permitAll()
-                .antMatchers("/registration", "/password_reset", "/password_reset_token").permitAll()
-                .antMatchers("/password_update").hasAuthority("CHANGE_PASSWORD_PRIVILEGE")
-                .antMatchers("/admin/**").hasAnyRole("ADMIN", "PREVIOUS_ADMINISTRATOR")
-                .anyRequest().authenticated()
+                .logout(logoutCustomizer -> logoutCustomizer
+                        .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
+                        .logoutSuccessUrl("/"))
 
-                .and().csrf().disable()
-                .formLogin()
-                .loginPage("/login").failureUrl("/?login=error")
-                .defaultSuccessUrl("/profile")
-                .usernameParameter("email")
-                .passwordParameter("password")
+                .exceptionHandling(exceptionCustomizer -> exceptionCustomizer
+                        .accessDeniedPage("/access-denied"))
 
-                .and()
-                .logout()
-                .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-                .logoutSuccessUrl("/").and().exceptionHandling()
-                .accessDeniedPage("/access-denied");
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("favicon.ico", "/resources/**", "/static/**", "/css/**", "/js/**", "/images/**").permitAll()
+                        .requestMatchers("/", "/login", "/profile/**", "/test").permitAll()
+                        .requestMatchers("/imprint").permitAll()
+                        .requestMatchers("/registration", "/password_reset", "/password_reset_token").permitAll()
+                        .requestMatchers("/password_update").hasAuthority("CHANGE_PASSWORD_PRIVILEGE")
+                        .requestMatchers("/admin/**").hasAnyRole("ADMIN", "PREVIOUS_ADMINISTRATOR")
+                        .anyRequest().authenticated())
+                .build();
     }
 
-    @Override
-    public void configure(WebSecurity web) {
-        web
-                .ignoring()
-                .antMatchers("/resources/**", "/static/**", "/css/**", "/js/**", "/images/**");
+    @Bean
+    public AuthenticationProvider authenticationProvider() {
+        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
+        authenticationProvider.setUserDetailsService(userDetailsService);
+        authenticationProvider.setPasswordEncoder(bCryptPasswordEncoder);
+        return authenticationProvider;
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
+        return config.getAuthenticationManager();
     }
 
     @Bean
@@ -91,3 +90,4 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
         return filter;
     }
 }
+
