@@ -3,34 +3,12 @@ package de.sremer.crawlicious.configuration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.servlet.ViewResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.view.AbstractUrlBasedView;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
 
 @Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
+public class WebMvcConfig {
 
     @Bean
     public BCryptPasswordEncoder passwordEncoder() {
-        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
-        return bCryptPasswordEncoder;
-    }
-
-    private ViewResolver getViewResolver() {
-        InternalResourceViewResolver viewResolver = new InternalResourceViewResolver() {
-            @Override
-            protected AbstractUrlBasedView buildView(String viewName) throws Exception {
-                String suffix = getSuffix();
-                if (suffix != null && suffix.length() > 0 && viewName.endsWith(suffix)) {
-                    viewName = viewName.substring(0, viewName.length() - suffix.length());
-                }
-
-                return super.buildView(viewName);
-            }
-        };
-        viewResolver.setPrefix("/");
-        viewResolver.setSuffix(".html");
-        return viewResolver;
+        return new BCryptPasswordEncoder();
     }
 }
