@@ -1,13 +1,15 @@
 package de.sremer.crawlicious.controller;
 
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
-@Controller
+@RestController
+@RequestMapping("/imprint")
 public class ImprintController {
 
-    @GetMapping(value = {"/imprint"})
+    @GetMapping
     public ModelAndView index() {
         return new ModelAndView("imprint");
     }
