@@ -1,7 +1,7 @@
 package de.sremer.crawlicious.controller;
 
-import de.sremer.crawlicious.service.external.TitleResolverService;
-import org.springframework.beans.factory.annotation.Autowired;
+import de.sremer.crawlicious.service.external.TitleResolver;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,17 +10,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/title")
+@RequiredArgsConstructor
 public class TitleResolverController {
 
-    private TitleResolverService titleResolverService;
-
-    @Autowired
-    TitleResolverController(TitleResolverService titleResolverService) {
-        this.titleResolverService = titleResolverService;
-    }
+    private final TitleResolver titleResolver;
 
     @GetMapping(value = "/resolve", produces = MediaType.TEXT_PLAIN_VALUE)
-    public String resolve(@RequestParam(value = "url", required = true) String url) {
-        return titleResolverService.getWebsiteTitle(url);
+    public String resolve(@RequestParam(value = "url") String url) {
+        return titleResolver.getWebsiteTitle(url);
     }
 }
